rebar: add alias for elvis
Overload the "eunit" target to run elvis using a hook. Add the alias
rebar plugin to add new targets.

The eunit target is overloaded because:

* rebar3 does not seem to have support for running external commands
  except using a pre/post_hook

* new targets can't be added except by making a new plugin

* not all targets call the pre/post_hooks

  For example, compile, clean, eunit, edoc and ct use hooks.

  dialyzer and cover do not use the hooks.

* the eunit command is not used (tests use common test)

The alternative is to use the rebar_lint plugin. Since we're already
downloading a plugin. It also has the advantage that it does not require
having elvis pre-installed.